Refactor problems in separate files, independent module interface
