+++ torchrl/problems/__init__.py
@@ -1,0 +1,5 @@
+from .a2c import A2CProblem
+from .ddpg import DDPGProblem
+from .dqn import DQNProblem
+from .ppo import PPOProblem
+from .prioritized_dqn import PrioritizedDQNProblem
