+++ torchrl/registry/__init__.py
@@ -1,3 +1,2 @@
 from .registry import *  # pylint: disable=wildcard-import
-from . import hparams
-from . import problems
+from .problems import Problem, HParams
